Squashed all changes and rebased. Addressed review comments.
